requirements-dev.txt
PyYAML
PyNaCl
pytest
pytest-cov
pre-commit
ruff
black
flake8
pylint
coverage
